+++ packages/api-reference/src/components/Content/Tags/TagSection.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { ScalarMarkdown } from '@scalar/components'
+import type { TraversedTag } from '@scalar/workspace-store/schemas/v3.1/strict/openapi-document'
 
 import { Anchor } from '@/components/Anchor'
 import { OperationsList } from '@/components/OperationsList'
@@ -13,7 +14,6 @@ import {
   SectionHeaderTag,
 } from '@/components/Section'
 import { SpecificationExtension } from '@/features/specification-extension'
-import type { TraversedTag } from '@/features/traverse-schema'
 import { useConfig } from '@/hooks/useConfig'
 
 const { tag, headerId, isCollapsed } = defineProps<{
@@ -45,7 +45,7 @@ const config = useConfig()
         <SectionColumn>
           <ScalarMarkdown
             :clamp="isCollapsed ? 7 : undefined"
-            :value="tag.tag?.description ?? ''"
+            :value="tag?.description ?? ''"
             withImages />
         </SectionColumn>
         <SectionColumn>
@@ -53,6 +53,6 @@ const config = useConfig()
         </SectionColumn>
       </SectionColumns>
     </SectionContent>
-    <SpecificationExtension :value="tag.tag" />
+    <SpecificationExtension :value="tag.xKeys" />
   </Section>
 </template>