+++ packages/json-magic/src/bundle/index.ts
@@ -1,3 +1,3 @@
-// biome-ignore lint/performance/noBarrelFile: <explanation>
-export { bundle } from './bundle'
-export type { Plugin, LoaderPlugin, LifecyclePlugin, ResolveResult } from './bundle'
+export type { LifecyclePlugin, LoaderPlugin, Plugin, ResolveResult } from './bundle'
+// biome-ignore lint/performance/noBarrelFile: exporting bundle
+export { bundle, resolveAndCopyReferences } from './bundle'