+++ packages/openapi-upgrader/src/index.ts
@@ -1,3 +1,2 @@
 // biome-ignore lint/performance/noBarrelFile: root entry point
-
 export { upgrade } from './upgrade'